AuthorizeControllers

+++ Api-Programacion3/Controllers/AdminController.cs
@@ -1,10 +1,14 @@
 ﻿using Application.Interfaces;
 using Application.Models.Dtos;
+using Application.Services;
 using Domain.Entities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace Api_Programacion3.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class AdminController : ControllerBase
@@ -19,33 +23,63 @@ namespace Api_Programacion3.Controllers
         [HttpGet("[action]")]
         public IActionResult GetAll()
         {
-            return Ok(_adminService.GetAdmins());
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "SysAdmin")
+            {
+                return Ok(_adminService.GetAdmins());
+            }
+            return Ok("Rol de usuario no calificado");
         }
 
         [HttpGet("[action]")]
         public IActionResult GetAdmin(string name)
         {
-            return Ok(_adminService.GetAdminByName(name));
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "SysAdmin")
+            {
+                return Ok(_adminService.GetAdminByName(name));
+            }
+            return Ok("Rol de usuario no calificado");
         }
 
         [HttpPost("[action]")]
         public IActionResult AddAmin([FromBody] AdminDto adminDto) 
         {
-            return Ok(_adminService.AddAdmin(adminDto));
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "SysAdmin")
+            {
+                return Ok(_adminService.AddAdmin(adminDto));
+            }
+            return Ok("Rol de usuario no calificado");
         }
 
         [HttpPut("[action]")]
         public IActionResult UpdateAdmin(int id,[FromBody] AdminDto adminDto)
         {
-            _adminService.UpdateAdmin(id, adminDto);
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "SysAdmin")
+            {
+                _adminService.UpdateAdmin(id, adminDto);
             return Ok("Admin actualizado");
+            }
+            return Ok("Rol de usuario no calificado");
         }
 
         [HttpDelete("[action]")]
         public IActionResult DeleteAdmin(int id)
-        { 
-            _adminService.DeleteAdmin(id);
+        {
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "SysAdmin")
+            {
+                _adminService.DeleteAdmin(id);
             return Ok("Admin eliminado");
+            }
+            return Ok("Rol de usuario no calificado");
         }
     }
 }
