+++ Api-Programacion3/Controllers/ItemController.cs
@@ -1,9 +1,13 @@
 ﻿using Application.Interfaces;
 using Application.Models.Dtos;
+using Application.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace Api_Programacion3.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class ItemController : ControllerBase
@@ -16,30 +20,61 @@ namespace Api_Programacion3.Controllers
         [HttpGet("[action]")]
         public IActionResult GetItemById(int id)
         {
-            return Ok(_itemService.GetItemById(id));
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "Client")
+            {
+                return Ok(_itemService.GetItemById(id));
+            }
+            return Ok("Rol de usuario no calificado"); 
         }
         [HttpGet("[action]")]
         public IActionResult GetAllItems()
         {
-            return Ok(_itemService.GetItems());
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "Admin")
+            {
+                 return Ok(_itemService.GetItems());
+            }
+            return Ok("Rol de usuario no calificado");
         }
         [HttpPost("[action]")]
         public IActionResult AddItem(ItemDto itemDto)
         {
-            return Ok(_itemService.AddItem(itemDto));
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "Client")
+            {
+                return Ok(_itemService.AddItem(itemDto));
+            }
+            return Ok("Rol de usuario no calificado");
+
+           
         }     
         [HttpPut("[action]")]
         public IActionResult UpdateItem(int id, int quantity)
         {
-            _itemService.UpdateItem(id, quantity);
-            return Ok("Item actualizado");
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
 
+            if (userRole == "Client")
+            {
+                _itemService.UpdateItem(id, quantity);
+                return Ok("Item actualizado");
+            }
+            return Ok("Rol de usuario no calificado");
         }
         [HttpDelete("[action]")]
         public IActionResult DeleteItem(int id)
         {
-            _itemService.DeleteItem(id);
-            return Ok("Item eliminado");
+            var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+
+            if (userRole == "Client")
+            {
+                _itemService.DeleteItem(id);
+                return Ok("Item eliminado");
+            }
+            return Ok("Rol de usuario no calificado");
         }
 
     }
